worker-service/requirements.txt
rq
redis
pychromecast
boto3
gtts
python-dotenv
requests
croniter
pytz
celery
